+++ src/base/Element.php
@@ -6018,14 +6018,15 @@ JS,
      *
      * If no partial template exists for the element, its string representation will be output instead.
      *
+     * @param array $variables
      * @return Markup
      * @throws InvalidConfigException
      * @throws NotSupportedException
      * @see ElementHelper::renderElements()
      * @since 5.0.0
      */
-    public function render(): Markup
+    public function render(array $variables = []): Markup
     {
-        return ElementHelper::renderElements([$this]);
+        return ElementHelper::renderElements([$this], $variables);
     }
 }