+++ src/elements/ElementCollection.php
@@ -76,14 +76,15 @@ class ElementCollection extends Collection
      *
      * If no partial template exists for an element, its string representation will be output instead.
      *
+     * @param array $variables
      * @return Markup
      * @throws InvalidConfigException
      * @throws NotSupportedException
      * @see ElementHelper::renderElements()
      * @since 5.0.0
      */
-    public function render(): Markup
+    public function render(array $variables = []): Markup
     {
-        return ElementHelper::renderElements($this->items);
+        return ElementHelper::renderElements($this->items, $variables);
     }
 }